Move Current Files to directories to setup for changes to come
