+++ Files/groupInfo.txt
@@ -1,0 +1,12 @@
+Welcome to Christian Student Union
+CSU is a place to connect, learn, serve and grow
+Hosted by: Pastor Serman Man
+Officers
+President: Fake Name
+Vice President: Not Real
+VP of Outreach: Really Cool Person
+Treasurer: Money Guy
+Secretary: Types Real Quick
+Representatives: John Smith, Joe Doe
+Worship Leader: Singer & Pianist
+Needed: Multimedia Director (PowerPoints, Fliers, etc.)
